% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{annotateSamples}
\alias{annotateSamples}
\title{annotate sample}
\usage{
annotateSamples(msexpori, FileSampleMapping, Samples2Remove = "")
}
\arguments{
\item{msexpori}{msexperiment}

\item{FileSampleMapping}{mapping of files to meaningfull label}

\item{Samples2Remove}{sample to remove}
}
\description{
annotate sample
}

