% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BXDFunctions.R
\name{createGraph}
\alias{createGraph}
\title{create an igraph}
\usage{
createGraph(adjMatrix, protnam)
}
\arguments{
\item{adjMatrix}{adjacency matrix}

\item{protnam}{protein names}
}
\description{
create an igraph
}

