% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{filterAndAggregate}
\alias{filterAndAggregate}
\title{filter ms exeperiment and aggreage peptides proteins}
\usage{
filterAndAggregate(exp, scorethresh = 0.005, n = 2,
  TOPPeptidesPerProtein = 5)
}
\arguments{
\item{exp}{- msexperiment}

\item{scorethresh}{- threshold for mscore}

\item{n}{- number of samples with mscore below scorethreshold}

\item{TOPPeptidesPerProtein}{- number of peptides per protein to use for quantification}
}
\description{
filter ms exeperiment and aggreage peptides proteins
}

