% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BXDFunctions.R
\name{graphMeasures}
\alias{graphMeasures}
\title{computes graph measures degree, closeness, betweeness}
\usage{
graphMeasures(gdiff)
}
\arguments{
\item{gdiff}{an igraph object}
}
\description{
computes graph measures degree, closeness, betweeness
}

