% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{keepSample}
\alias{keepSample}
\title{keep samples containing string}
\usage{
keepSample(exp, tiss = "Brain")
}
\arguments{
\item{exp}{ms experiment}

\item{tiss}{substring the sample must contain, it is colled tiss because it is the tissue name usually}
}
\description{
keep samples containing string
}

