% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BXDFunctions.R
\name{plotGraph}
\alias{plotGraph}
\title{plot igraph facade}
\usage{
plotGraph(matr, names, main = "cor(HDF)-cor(CD)", vertcol = NULL,
  repulserad = 10000, myl = NULL)
}
\arguments{
\item{matr}{an adjacency matrix}

\item{names}{protein names}

\item{main}{main title}

\item{vertcol}{color of vertices}

\item{repulserad}{repulsion}

\item{myl}{my layout - pass your own i graph layout}
}
\description{
plot igraph facade
}

