% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BXDFunctions.R
\name{plotGraph2}
\alias{plotGraph2}
\title{plot a graph - facade to igraphs plot function}
\usage{
plotGraph2(graph, myl, cex = 0.5, main = "", vertcol = NULL)
}
\arguments{
\item{graph}{an igraph}

\item{myl}{my layout}

\item{cex}{size of vertex label}

\item{main}{main title}

\item{vertcol}{vertex color}
}
\description{
plot a graph - facade to igraphs plot function
}

