% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BXDFunctions.R
\name{plotGraphAnn}
\alias{plotGraphAnn}
\title{plot a graph - facade to igraphs plot function}
\usage{
plotGraphAnn(graph, myl, vfc = "white", family = 1)
}
\arguments{
\item{graph}{an igraph}

\item{myl}{my layout}

\item{vfc}{vertex.frame.color}

\item{family}{vertex.label.family}
}
\description{
plot a graph - facade to igraphs plot function
}

