% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BXDFunctions.R
\name{plotGraphAnn2}
\alias{plotGraphAnn2}
\title{plot a graph - facade to igraphs plot function}
\usage{
plotGraphAnn2(graph, myl, vfc = "white", family = 1, edge.width = NULL,
  vertex.label.cex = 1, vertex.label.color = 1)
}
\arguments{
\item{graph}{an igraph}

\item{myl}{my layout}

\item{vfc}{vertex frame color}

\item{family}{vertex.label.family}

\item{edge.width}{edge.width}

\item{vertex.label.cex}{vertex.label.cex}

\item{vertex.label.color}{vertex.label.color}
}
\description{
plot a graph - facade to igraphs plot function
}

