% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BXDFunctions.R
\name{plotGraphSimple}
\alias{plotGraphSimple}
\title{plot a graph - facade to igraphs plot function}
\usage{
plotGraphSimple(graph, myl = NULL)
}
\arguments{
\item{graph}{an igraph}

\item{myl}{my layout}
}
\description{
plot a graph - facade to igraphs plot function
}

