% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BXDFunctions.R
\name{plotHistCor2Cond}
\alias{plotHistCor2Cond}
\title{plots histograms from upper left, lower left, and lower right quarter of matrix}
\usage{
plotHistCor2Cond(tmp, main = "")
}
\arguments{
\item{tmp}{matrix}

\item{main}{main title of figure}
}
\description{
plots histograms from upper left, lower left, and lower right quarter of matrix
}

