% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{prefilterNewDataSet}
\alias{prefilterNewDataSet}
\title{load eperiment, aggregate peptides from top transitions etc.}
\usage{
prefilterNewDataSet(SpecLib, maxNRTransitions = 3, maxNRPeptides = 10)
}
\arguments{
\item{SpecLib}{path to openswath output}

\item{maxNRTransitions}{max transitions per peptide}

\item{maxNRPeptides}{nr peptides per protein to keeps}
}
\description{
removes decoys
keeps only proteotypic peptides
}

