% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{selectMatchStrains}
\alias{selectMatchStrains}
\title{Selects those columns with same strain information}
\usage{
selectMatchStrains(all)
}
\arguments{
\item{all}{matrix starting wit columns named either Mito_ or Tissue_}
}
\description{
Selects those columns with same strain information
}

